==> 01_know_database.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import tabula
from astropy.coordinates import SkyCoord
from astropy import units as u

df_known = pd.DataFrame(columns=['Name', 'ra', 'dec', 'n'])

# Lemmon database

df_lemon = pd.read_csv('data/known_gls.csv')
df_lemon

df_known['Name'] = df_lemon['Name']
df_known['ra'] = df_lemon['RA']
df_known['dec'] = df_lemon['DEC'].astype(float)
df_known['n'] = df_lemon['N_images']
df_known

df_known.info()

"""
java -jar .venv/lib/python3.12/site-packages/tabula/tabula-1.0.5-jar-with-dependencies.jar -p 14 -a 100,10,410,600 -o ./data/strides_quads.csv "/Users/alexeysergeyev/Zotero/storage/GYY7NDJT/Schmidt et al. - 2023 - STRIDES automated uniform models for 30 quadruply imaged quasars.pdf"
"""
df_strides = pd.read_csv('data/strides_quads.csv', header=None)

df_strides.replace(' ', '', regex=True, inplace=True)
# Replace non-standard minus sign with standard minus sign
df_strides[2] = df_strides[2].astype('str').str.replace('−', '-')

# Convert the column to float
df_strides[1] = df_strides[1].astype(float)
df_strides[2] = df_strides[2].astype(float)
# df_strides[2].astype('str').replace('−', '-', regex=True, inplace=True)
df_strides

coords_known = SkyCoord(ra=df_known['ra'], dec=df_known['dec'], unit=(u.deg, u.deg))
coords_strides = SkyCoord(ra=df_strides[1], dec=df_strides[2], unit=(u.deg, u.deg))
idx, d2d, _, = coords_strides.match_to_catalog_sky(coords_known)
df_strides['sep'] = d2d.to(u.arcsec).value
cond = df_strides['sep'] > 10
df_add = df_strides.loc[cond, [0, 1, 2]]
# df_add['n'] = 4 * np.ones(df_add.shape[0])
df_add[1] = df_add[1].astype(float)
df_add[2] = df_add[2].astype(float)
df_add

df_known = pd.concat([df_known, df_add[[0, 1, 2]].rename(columns={0: 'Name', 1: 'ra', 2: 'dec'})], ignore_index=True)
df_known


df_known.info()

# 150 candidates

"""
java -jar .venv/lib/python3.12/site-packages/tabula/tabula-1.0.5-jar-with-dependencies.jar -p 4-5 -a 130,40,800,600 -o ./data/lemon150_1-2.csv "/Users/alexeysergeyev/Zotero/storage/BCFE69ZR/Lemon et al. - 2022 - Gravitationally lensed quasars in Gaia -- IV. 150 new lenses, quasar pairs, and projected quasars.pdf"
"""
"""
java -jar .venv/lib/python3.12/site-packages/tabula/tabula-1.0.5-jar-with-dependencies.jar -p 6 -a 120,40,660,600 -o ./data/lemon150_3.csv "/Users/alexeysergeyev/Zotero/storage/BCFE69ZR/Lemon et al. - 2022 - Gravitationally lensed quasars in Gaia -- IV. 150 new lenses, quasar pairs, and projected quasars.pdf"
"""

df1 = pd.read_csv('data/lemon150_1-2.csv')
df2 = pd.read_csv('data/lemon150_3.csv')
df1.dropna(subset=['Name'], inplace=True)
df2.dropna(subset=['Name'], inplace=True)
df_lemon150 = pd.concat([df1, df2])
df_lemon150

cond = df_lemon150['Classification'].str.contains('lens')
df_lens = df_lemon150[cond]
df_lens['Dec.'].astype('str').replace('−', '-', regex=True, inplace=True)
df_lens.loc[:, 'Name'] = df_lens['Name'].astype('str')
df_lens

df_lens.info()

coords_known = SkyCoord(ra=df_known['ra'], dec=df_known['dec'], unit=(u.deg, u.deg))
coords = SkyCoord(ra=df_lens['R.A.'], dec=df_lens['Dec.'], unit=(u.deg, u.deg))
idx, d2d, _, = coords.match_to_catalog_sky(coords_known)
df_lens.loc[:, 'sep'] = d2d.to(u.arcsec).value
cond = df_lens['sep'] > 10
df_add = df_lens[cond].reset_index(drop=True)
# df_add['n'] = 4
df_add

df_known = pd.concat([df_known, df_add[['Name', 'R.A.', 'Dec.']].rename(columns={'R.A.': 'ra', 'Dec.': 'dec'})], ignore_index=True)
df_known

df_known.info()

df_known.to_csv('data/my_gls.csv', index=False)

df_known = pd.read_csv('data/my_gls.csv')
df_known
