==> 02_1_checking.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from matplotlib.patches import Ellipse
import os

scan = os.scandir('./data/candidates/lists/')
files = [f.path for f in scan if f.is_file() and f.name.endswith('_2.csv')]
files.sort()
files

df_candidates = None
for file in files[:]:
    df = pd.read_csv(file)
    if df_candidates is None:
        df_candidates = df
    else:
        df_candidates = pd.concat([df_candidates, df])
df_candidates

df_candidates.to_csv('./data/candidates/candidates2.csv', index=False)

df_candidates = pd.read_csv('./data/candidates/candidates2.csv')
gr = df_candidates.groupby('source_id_smss')
len(gr)
